+++ AndroidXCI/lib/src/main/kotlin/dev/androidx/ci/gcloud/GoogleCloudApi.kt
@@ -69,7 +69,7 @@ internal interface GoogleCloudApi {
     /**
      * Walks all entries in the given gcsPath from top to bottom.
      */
-    suspend fun walkTopDown(gcsPath: GcsPath): Sequence<BlobVisitor>
+    suspend fun walkEntires(gcsPath: GcsPath): Sequence<BlobVisitor>
 }
 
 /**
@@ -83,7 +83,7 @@ internal suspend fun GoogleCloudApi.download(
     target: File,
     filter: (String) -> Boolean
 ) {
-    walkTopDown(gcsPath).filter { visitor ->
+    walkEntires(gcsPath).filter { visitor ->
         filter(visitor.relativePath)
     }.forEach { visitor ->
         val targetFile = if (visitor.isRoot()) {
@@ -138,7 +138,7 @@ private class GoogleCloudApiImpl(
         GcsPath.create(blob)
     }
 
-    override suspend fun walkTopDown(
+    override suspend fun walkEntires(
         gcsPath: GcsPath
     ): Sequence<BlobVisitor> {
         val blobId = gcsPath.blobId
@@ -200,14 +200,34 @@ private class GoogleCloudApiImpl(
         "${config.bucketPath}/$relativePath"
 }
 
+/**
+ * Provides access to a Blob returned from a [GoogleCloudApi.walkEntires] method.
+ */
 internal interface BlobVisitor {
+    /**
+     * Returns true if this Blob is the root blob that matches the `gcsPath` parameter of [GoogleCloudApi.walkEntires].
+     */
     fun isRoot() = relativePath.isEmpty()
 
+    /**
+     * Returns the relative path of the blob wrt to the `gcsPath` parameter of [GoogleCloudApi.walkEntires].
+     */
     val relativePath: String
+
+    /**
+     * Full gcsPath of the blob
+     */
     val gcsPath: GcsPath
 
+    /**
+     * The filename of the blob (name after the last `/` in the [gcsPath].
+     */
     val fileName: String
         get() = gcsPath.path.substringAfterLast('/')
+
+    /**
+     * Opens the input stream to the blob. You must make sure to close it after using it.
+     */
     fun obtainInputStream(): InputStream
 }
